Fixes for CentOS 6/7.
CentOS 6 and 7 have different paths where mailcatcher is
installed.
Updated dependencies version numbers.
Add rubygem-json_pure as it seems to be a runtime dependency
for mailcatcher in CentOS7 which is not installed by default.
Add rubygem-mime-types as it requires ruby >= 1.9.2 now which
is not available in CentOS6.
Maybe you need the epel repository so that these packages get
installed.