Fixing CentOS6 again.
This gets really messy, but newer versions of mailcatcher have dependencies on
gems which depend on gems that are not compatible with CentOS6's ruby version
1.8.7 anymore. To get a working mailcatcher on CentOS6 I need to stick with
an older mailcatcher 0.5.12. This requires a specific compatible i18n gem and
a specific eventmachine gem, as newer version do not accept mails in background
mode anymore. This is fixed in current mailcatcher version, but that one is
not compatible because of the other dependencies. WTF :-!